Update snackbar duration and clean up imports in navigation graph

+++ app/src/main/java/com/capella/screens/MoodCheckIn.kt
@@ -165,7 +165,7 @@ fun MoodSelectionScreen(
 
                         snackbarHostState.showSnackbar(
                             message = " Your mood '${it.icon}' has been recorded!",
-                            duration = SnackbarDuration.Short,
+                            duration = SnackbarDuration.Long,
                         )
                     }
                     //toast message
